+++ banner-rotation/model/statistics.go
@@ -2,8 +2,8 @@ package model
 
 //Statistics represents statistics
 type Statistics struct {
-	BannerID       int64  `json:"bannerId"`
-	GroupID        int64  `json:"groupId"`
-	NumberOfShows  uint32 `json:"numberOfShows"`
-	NumberOfClicks uint32 `json:"NumberOfClicks"`
+	BannerID       int64 `json:"bannerId"`
+	GroupID        int64 `json:"groupId"`
+	NumberOfShows  int64 `json:"numberOfShows"`
+	NumberOfClicks int64 `json:"NumberOfClicks"`
 }
